updates for eth 1.5.5 and cleanup
