--- clients/client3.go
@@ -1,19 +1,0 @@
-package main
-
-import (
-    "fmt"
-    )
-
-func main() {
-    fmt.Println("Starting client3")
-
-    a := "0x0"
-    
-    if a != "0x0" {
-        fmt.Println("not equal")
-    } else {
-        fmt.Println("equal")
-    }
-
-    fmt.Println("Terminating client3")
-}
