--- clients/deregister/Makefile
@@ -1,4 +1,0 @@
-# deregister make
-
-default:
-	go install
