--- clients/device/Makefile
@@ -1,4 +1,0 @@
-# bootstrap make
-
-default:
-	go install
