--- clients/rest_container_provider/Makefile
@@ -1,4 +1,0 @@
-# rest container provider make
-
-default:
-	go install
