--- contracts/whisper_directory.json
@@ -1,1 +1,0 @@
-{"info":{"language":"Solidity","languageVersion":"0.3.1","abiDefinition":[{"inputs":[],"type":"function","constant":false,"name":"kill","outputs":[]},{"inputs":[],"type":"function","constant":false,"name":"delete_entry","outputs":[{"type":"bool","name":"r"}]},{"inputs":[{"type":"address","name":"_account"}],"type":"function","constant":false,"name":"delete_entry_admin","outputs":[{"type":"bool","name":"r"}]},{"inputs":[{"type":"string","name":"_whisper_account"}],"type":"function","constant":false,"name":"add_entry","outputs":[{"type":"bool","name":"r"}]},{"inputs":[{"type":"address","name":"_account"}],"type":"function","constant":true,"name":"get_entry","outputs":[{"type":"string","name":"r"}]},{"inputs":[],"type":"constructor","constant":false,"name":"","outputs":null},{"inputs":[{"type":"uint256","name":"_eventcode"},{"type":"address","name":"_adder"},{"type":"string","name":"_whisper_account"}],"type":"event","constant":false,"name":"AddEntry","outputs":null},{"inputs":[{"type":"uint256","name":"_eventcode"},{"type":"address","name":"_deleter"},{"type":"string","name":"_whisper_account"}],"type":"event","constant":false,"name":"DeleteEntry","outputs":null}],"compilerVersion":"0.3.1","developerDoc":{"methods":{}},"userDoc":{"methods":{}},"source":"contract whisper_directory {address owner;mapping (address =\u003e string) whisper_accounts;enum event_codes {add_entry_event_code,delete_entry_event_code,admin_delete_entry_event_code}event AddEntry(uint indexed _eventcode, address indexed _adder, string _whisper_account) anonymous;event DeleteEntry(uint indexed _eventcode, address indexed _deleter, string _whisper_account) anonymous;function whisper_directory() {owner = msg.sender;}function add_entry(string _whisper_account) returns (bool r) {whisper_accounts[tx.origin] = _whisper_account;AddEntry(uint(event_codes.add_entry_event_code), tx.origin, _whisper_account);return true;}function get_entry(address _account) constant returns (string r) {return whisper_accounts[_account];}function delete_entry() returns (bool r) {var deleted_wa = whisper_accounts[tx.origin];DeleteEntry(uint(event_codes.delete_entry_event_code), tx.origin, deleted_wa);delete whisper_accounts[tx.origin];return true;}function delete_entry_admin(address _account) returns (bool r) {if (tx.origin == owner) {var deleted_wa = whisper_accounts[_account];DeleteEntry(uint(event_codes.admin_delete_entry_event_code), _account, deleted_wa);delete whisper_accounts[_account];return true;} else {return false;}}function kill() {if (msg.sender == owner) suicide(owner);}}"},"code":"0x606060405260008054600160a060020a03191633179055610569806100246000396000f3606060405260e060020a600035046341c0e1b5811461004757806367f4f2a61461006f5780637406d4df146100de5780638bcfa7d314610168578063c71da7d014610233575b005b61004560005433600160a060020a039081169116141561056757600054600160a060020a0316ff5b32600160a060020a03166000818152600160208181526040808420815183815281546002818716156101000260001901909116049381018490526102b89691949193919285928291820190849080156104185780601f106103ed57610100808354040283529160200191610418565b6102b8600435600080548190600160a060020a039081163290911614156104c05750600160a060020a03821680825260016020818152604080852081518381528154600295811615610100026000190116859004938101849052909493928592829190820190849080156104f65780601f106104cb576101008083540402835291602001916104f6565b6040805160206004803580820135601f81018490048402850184019095528484526102b89491936024939092918401919081908401838280828437509496505050505050506000816001600050600032600160a060020a031681526020019081526020016000206000509080519060200190828054600181600116156101000203166002900490600052602060002090601f016020900481019282601f106103bd57805160ff19168380011785555b506103389291505b80821115610467576000815560010161021f565b60408051602081810183526000808352600160a060020a03600435908116825260018084529185902080548651600294821615610100026000190190911693909304601f81018590048502840185019096528583526102ca959194939091908301828280156104965780601f1061046b57610100808354040283529160200191610496565b60408051918252519081900360200190f35b60405180806020018281038252838181518152602001915080519060200190808383829060006004602084601f0104600f02600301f150905090810190601f16801561032a5780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b505032600160a060020a031660008360405180806020018281038252838181518152602001915080519060200190808383829060006004602084601f0104600f02600301f150905090810190601f1680156103a75780820380516001836020036101000a031916815260200191505b509250505060405180910390a25060015b919050565b82800160010185558215610217579182015b828111156102175782518260005055916020019190600101906103cf565b820191906000526020600020905b8154815290600101906020018083116103fb57829003601f168201915b50509250505060405180910390a232600160a060020a03166000908152600160208190526040822080549281559160029181161561010002600019011604601f8190106104a257505b50600191505b5090565b820191906000526020600020905b81548152906001019060200180831161047957829003601f168201915b505050505090506103b8565b601f016020900490600052602060002090810190610461919061021f565b600091505b50919050565b820191906000526020600020905b8154815290600101906020018083116104d957829003601f168201915b50509250505060405180910390a2600160a060020a0383166000908152600160208190526040822080549281559160029181161561010002600019011604601f81901061054957505b50600191506104c5565b601f01602090049060005260206000209081019061053f919061021f565b56"}
