+++ integration/Makefile
@@ -9,7 +9,7 @@ containers:
 
 compose/test.yml: tmpl/contest.yml
 	mkdir -p compose
-	./buildtest.sh $(PASSWD)
+	cat tmpl/contest.yml >>compose/test.yml
 
 starttest: compose/test.yml
 	docker-compose -f compose/test.yml up -d
@@ -18,7 +18,7 @@ endtest:
 	docker-compose -f compose/test.yml stop
 	docker-compose -f compose/test.yml rm -f
 
-testdev: containers compose/test.yml starttest
+runtest: containers compose/test.yml starttest
 	docker-compose -f compose/test.yml logs
 
 compilecontracts:
@@ -28,11 +28,7 @@ compilecontracts:
 	docker-compose -f compose/test.yml up -d
 
 refreshJSON:
-	docker cp compose_compilecon_1:/root/container_executor.json ../contracts/
-	docker cp compose_compilecon_1:/root/device_registry.json ../contracts/
 	docker cp compose_compilecon_1:/root/directory.json ../contracts/
-	docker cp compose_compilecon_1:/root/token_bank.json ../contracts/
-	docker cp compose_compilecon_1:/root/whisper_directory.json ../contracts/
 	docker cp compose_compilecon_1:/root/agreements.json ../contracts/
 
 cleanslate: endtest clean
